Refactor libtock
Closes #143

  * Move firestorm functionality into separate libs (spi/nrf51/timer)

  * Fixup example apps

  * Fixed a few lint issues
