+++ userland/examples/c_hello/main.c
@@ -5,13 +5,14 @@
 #include <stdlib.h>
 #include <unistd.h>
 
-#include <firestorm.h>
+#include <console.h>
 
 char hello[] = "Hello World!\r\n";
 
-void nop(int x, int y, int z, void *ud) {}
+void nop() {}
 
-void main() {
+int main() {
   putnstr_async(hello, sizeof(hello), nop, NULL);
+  return 0;
 }
