+++ userland/libtock/console.h
@@ -1,11 +1,5 @@
-/* tock_str
- *
- * Lightweight string maniuplation and printing for Tock
- *
- */
-
-#ifndef TOCK_STR_H
-#define TOCK_STR_H
+#ifndef CONSOLE_H
+#define CONSOLE_H
 
 #include "tock.h"
 
@@ -13,4 +7,4 @@ void putstr(const char* str);
 void putnstr(const char* str, size_t len);
 void putnstr_async(const char* str, size_t len, subscribe_cb cb, void* userdata);
 
-#endif // TOCK_STR_H
+#endif // CONSOLE_H
